requirements.txt
pyyaml==6.0.1
numpy==1.24.3
scikit-learn==1.3.0
pingouin==0.5.3
typing-extensions==4.8.0 